agent/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
distro==1.9.0
filelock==3.18.0
fuzzywuzzy==0.18.0
google-ai-generativelanguage==0.4.0
google-api-core==2.24.2
google-auth==2.38.0
google-generativeai==0.3.1
googleapis-common-protos==1.69.2
greenlet==3.1.1
griffe==1.7.2
grpcio==1.71.0
grpcio-status==1.62.3
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
jiter==0.9.0
mcp==1.6.0
openai==1.70.0
openai-agents==0.0.8
playwright==1.51.0
proto-plus==1.26.1
protobuf==4.25.6
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.2
pydantic-settings==2.8.1
pydantic_core==2.33.1
pyee==12.1.1
python-dotenv==1.1.0
PyYAML==6.0.2
requests==2.32.3
rsa==4.9
sniffio==1.3.1
sse-starlette==2.2.1
starlette==0.46.1
tqdm==4.67.1
types-requests==2.32.0.20250328
typing-inspection==0.4.0
typing_extensions==4.13.1
urllib3==2.3.0
uvicorn==0.34.0
